fix: donot crash chain

+++ test/e2e/tests/xcall.go
@@ -66,7 +66,7 @@ func (x *XCallTestSuite) TestRollback(ctx context.Context, t *testing.T, chainA,
 	height, err := chainA.(ibc.Chain).Height(ctx)
 	x.Require().NoError(err)
 	ctx, err = chainB.ExecuteCall(ctx, res.RequestID, res.Data)
-	code, err := chainA.FindCallResponse(ctx, int64(height), res.SerialNo)
+	code, err := chainA.FindCallResponse(ctx, height, res.SerialNo)
 	x.Require().NoError(err)
 	assert.Equal(t, "0", code)
 	ctx, err = chainA.ExecuteRollback(ctx, res.SerialNo)
